DEMO: soffice: Announce coordinates of selected cells (#6897)
This is just a demo to show how handling the
different 'event_selection*' events for
SymphonyTable and SymphonyTableCell could in general
be used to announce the current selection of multiple
cells.

It adds a method 'announce_selected_cells' to
SymphonyTable that just announces the names
of the first and last cell in the current selection
for now.

To avoid the same selection being announced multiple
times, the first and last cell of the last time
the method has been called is remembered and the
announcement is only done if the current selection
is different.
(The method may be called multiple times for the same
selection, since e.g. multiple SELECTION_ADD events
are triggerdd when multiple cells are added to a
selection at the same time.)

NOTES:

* This demo requires a current development version of
  LibreOffice to work.

* 'event_selectionRemove' is not triggered as expected,
  so the selection is not announced when removing cells
  from the selection using Shift+arrow keys.
  This looks like a problem on LibreOffice side.

* There are some other cases where events are not
  triggered as expeced on LO side, s.e.g.
  tdf#118508 [1] or tdf#18748 [2], which are
  also LibreOffice problems that need be handled there.

(Note: Strictly speaking, the change to
SymphonyTable::getSelectedItemsCount is unrelated, just makes
use of the newly added implementation of IAccessibleTable2.)

[1] https://bugs.documentfoundation.org/show_bug.cgi?id=118508
[2] https://bugs.documentfoundation.org/show_bug.cgi?id=118748